Оптимизация кода: слабые ссылки, налажено взаимодействие между контроллером и фабрикой
