+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,14 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Наталья Черномырдина on 29.01.2025.
+//
+
+import Foundation
+
+struct QuizQuestion { // структура вопроса
+    let image: String // название картинки афиши/фильма
+    let text: String // вопрос
+    let correctAnswer: Bool // правильный ответ на вопрос
+}
