+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Наталья Черномырдина on 29.01.2025.
+//
+
+import Foundation
+
+struct QuizResultsViewModel { // структура вью модели "Результат квиза"
+    let title: String // заголовок алерта
+    let text: String  // количество набранных очков
+    let buttonText: String // текст конпки алерта
+}
