+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Наталья Черномырдина on 29.01.2025.
+//
+
+import UIKit
+
+struct QuizStepViewModel { // структура вью модели "Вопрос показан"
+    let image: UIImage // картинка
+    let question: String // вопрос
+    let questionNumber: String // счетчик вопросов
+}
+
